pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.39",
    "werkzeug>=3.1.3",
    "reportlab>=4.3.1",
    "requests>=2.32.3",
    "flask-wtf>=1.2.2",
    "wtforms>=3.2.1",
    "pandas>=2.2.3",
    "openpyxl>=3.1.5",
    "twilio>=9.5.1",
    "sendgrid>=6.11.0",
    "flask-mail>=0.10.0",
    "python-dotenv>=1.0.1",
    "qrcode>=8.0",
    "pillow>=11.1.0",
    "markupsafe>=3.0.2",
]
